cmt3d: change derivative extensions to Mrr,Mtt,Mpp,Mrt,Mrp,Mtp,dep,lon,lat,ctm,hdr,        set npts1=0 before reading derivative syn to make sure they do exist grid3d: eliminate the use of real*8 mech_table.pl: fix the case when CMTSOLUTION_GRD does not exist

+++ grid3d/grid3d_variables.f90
@@ -11,7 +11,7 @@ module grid3d_variables
   ! data selection
   character(len=150) :: flexwin_out_file
   logical :: weigh_data_files
-  real*8 :: comp_z_weight, comp_r_weight, comp_t_weight,&
+  real :: comp_z_weight, comp_r_weight, comp_t_weight,&
        az_exp_weight, &
        pnl_dist_weight, rayleigh_dist_weight, love_dist_weight
